refact: factor out FileDisplay component from drop box to be shared
