+++ src/components/FileDisplay.js
@@ -1,0 +1,194 @@
+import React, { useCallback, useEffect, useMemo, useState } from "react";
+import { useAuthorizationHeader } from "../lib/auth/utils";
+import { Alert, Spin } from "antd";
+
+import fetch from "cross-fetch";
+
+import { Document, Page } from "react-pdf/dist/esm/entry.webpack5";
+
+import { Light as SyntaxHighlighter } from "react-syntax-highlighter";
+import { a11yLight } from "react-syntax-highlighter/dist/cjs/styles/hljs";
+import {
+    bash,
+    dockerfile,
+    json,
+    markdown,
+    plaintext,
+    python,
+    r,
+    shell,
+    xml
+} from "react-syntax-highlighter/dist/cjs/languages/hljs";
+
+import JsonDisplay from "./JsonDisplay";
+
+import "react-pdf/dist/esm/Page/AnnotationLayer.css";
+import "react-pdf/dist/esm/Page/TextLayer.css";
+import PropTypes from "prop-types";
+
+SyntaxHighlighter.registerLanguage("bash", bash);
+SyntaxHighlighter.registerLanguage("dockerfile", dockerfile);
+SyntaxHighlighter.registerLanguage("json", json);
+SyntaxHighlighter.registerLanguage("markdown", markdown);
+SyntaxHighlighter.registerLanguage("plaintext", plaintext);
+SyntaxHighlighter.registerLanguage("python", python);
+SyntaxHighlighter.registerLanguage("r", r);
+SyntaxHighlighter.registerLanguage("shell", shell);
+SyntaxHighlighter.registerLanguage("xml", xml);
+
+const BASE_PDF_OPTIONS = {
+    cMapUrl: "cmaps/",
+    cMapPacked: true,
+    standardFontDataUrl: "standard_fonts/",
+};
+
+const LANGUAGE_HIGHLIGHTERS = {
+    ".bash": "bash",
+    ".json": "json",
+    ".md": "markdown",
+    ".txt": "plaintext",
+    ".py": "python",
+    ".R": "r",
+    ".sh": "shell",
+    ".xml": "xml",
+
+    // Special files
+    "Dockerfile": "dockerfile",
+    "README": "plaintext",
+    "CHANGELOG": "plaintext",
+};
+
+export const VIEWABLE_FILE_EXTENSIONS = [...Object.keys(LANGUAGE_HIGHLIGHTERS), ".pdf"];
+
+const FileDisplay = ({ uri, fileName, loading }) => {
+    const authHeader = useAuthorizationHeader();
+
+    const [fileLoadError, setFileLoadError] = useState("");
+    const [loadingFileContents, setLoadingFileContents] = useState(false);
+    const [fileContents, setFileContents] = useState({});
+    const [pdfPageCounts, setPdfPageCounts] = useState({});
+
+    const pdfOptions = useMemo(() => ({
+        ...BASE_PDF_OPTIONS,
+        httpHeaders: authHeader,
+    }), [authHeader]);
+
+    let textFormat = false;
+    if (fileName) {
+        Object.keys(LANGUAGE_HIGHLIGHTERS).forEach(ext => {
+            if (fileName.endsWith(ext)) {
+                textFormat = true;
+            }
+        });
+    }
+
+    const fileExt = fileName ? fileName.split(".").slice(-1)[0] : null;
+
+    useEffect(() => {
+        // File changed, so reset the load error
+        setFileLoadError("");
+
+        (async () => {
+            if (!fileName) return;
+
+            if (fileExt === "pdf") {
+                setLoadingFileContents(true);
+            }
+
+            if (!textFormat || fileContents.hasOwnProperty(uri)) return;
+
+            if (!uri) {
+                console.error(`Files: something went wrong while trying to load ${uri}`);
+                setFileLoadError("Could not find URI for file.");
+                return;
+            }
+
+            try {
+                setLoadingFileContents(true);
+                const r = await fetch(uri, { headers: authHeader });
+                if (r.ok) {
+                    const text = await r.text();
+                    const content = (fileExt === "json" ? JSON.parse(text) : text);
+                    setFileContents({
+                        ...fileContents,
+                        [uri]: content,
+                    });
+                } else {
+                    setFileLoadError(`Could not load file: ${r.content}`);
+                }
+            } catch (e) {
+                console.error(e);
+                setFileLoadError(`Could not load file: ${e.message}`);
+            } finally {
+                setLoadingFileContents(false);
+            }
+        })();
+    }, [uri]);
+
+    const onPdfLoad = useCallback(({numPages}) => {
+        setLoadingFileContents(false);
+        setPdfPageCounts({...pdfPageCounts, [uri]: numPages});
+    }, [uri]);
+
+    const onPdfFail = useCallback(err => {
+        console.error(err);
+        setLoadingFileContents(false);
+        setFileLoadError(`Error loading PDF: ${err.message}`);
+    }, []);
+
+    if (!uri || !fileName) {
+        return <div />;
+    }
+
+    return <Spin spinning={loading || loadingFileContents}>
+        {(() => {
+            if (fileLoadError) {
+                return <Alert
+                    type="error"
+                    message={`Error loading file: ${file}`}
+                    description={fileLoadError}
+                />;
+            }
+
+            if (fileExt === "pdf") {  // Non-text, content isn't loaded a priori
+                return (
+                    <Document file={uri} onLoadSuccess={onPdfLoad} onLoadError={onPdfFail} options={pdfOptions}>
+                        {(() => {
+                            const pages = [];
+                            for (let i = 1; i <= pdfPageCounts[file] ?? 1; i++) {
+                                pages.push(<Page pageNumber={i} key={i} />);
+                            }
+                            return pages;
+                        })()}
+                    </Document>
+                );
+            } else if (fileExt === "json") {
+                const jsonSrc = fileContents[uri];
+                if (loadingFileContents || !jsonSrc) {
+                    return <div/>;
+                }
+                return (
+                    <JsonDisplay jsonSrc={jsonSrc}/>
+                );
+            } else {  // if (textFormat)
+                return (
+                    <SyntaxHighlighter
+                        language={LANGUAGE_HIGHLIGHTERS[`.${fileExt}`]}
+                        style={a11yLight}
+                        customStyle={{fontSize: "12px"}}
+                        showLineNumbers={true}
+                    >
+                        {fileContents[uri] || ""}
+                    </SyntaxHighlighter>
+                );
+            }
+        })()}
+    </Spin>;
+};
+FileDisplay.propTypes = {
+    uri: PropTypes.string,
+    fileName: PropTypes.string,
+    loading: PropTypes.bool,
+};
+
+export default FileDisplay;
